Move NoopWebhook to Util
